Fix nginx configuration issues with how we can trigger canarytokens
Summary:
We noticed that tokens were not being triggered unless there was
a . in the url path after the hostname. This happened because we
were sending those requests to the frontend process and not the
switchboard. We figured it was a nginx configuration issue.

Test Plan:
Spin up a local docker with new nginx conf. Check that we are alerted
on urls without a . in it. Subsequently check that normal token trigger
methods work as well as normal navigation.

Reviewers: #devs

Differential Revision: http://phabricator.thinkst.com:9000/D956