Prevent certbot from upgrading and creating redirects
Summary:
With a latest change to certbot-auto, it would alter our `nginx.conf` file
to the point that it would 404 non-ssl traffic. This is required for our http
tokens as we don't enforce https on them.

Adding the `--no-redirect` flag tells certbot-auto to not create the redirects.

In addition, we've added the `--no-self-upgrade` flag so that we don't get suprise
breakages going forward.

Test Plan: We tested this while fixing the live instance.

Reviewers: azhar

Reviewed By: azhar

Differential Revision: http://phabricator.thinkst.com:9000/D2309