Rename .env files to .env.dist
Summary:
Since we manually edit the `.env` files to configure the server, it
makes sense to rename the distributed `.env` files and mark them as `.env.dist`
to ensure that pulling new code from the repo doesn't mistakenly blow away
local config.

Test Plan:
Ensured that the README file explains the change as well as includes
the step for setup.

Reviewers: azhar

Reviewed By: azhar

Differential Revision: http://phabricator.thinkst.com:9000/D2310